+++ lib/utils/jest.setup.js
@@ -1,6 +1,6 @@
 import { NullConsole } from '@jest/console'
 import { transports } from 'winston'
-import logger from '%src/lib/logger'
+import logger from '%src/tools/logger'
 
 const getConsoleTransport = () => logger.transports.find(
 	(transport) => transport.constructor.name === transports.Console.name,
