+++ src/index.js
@@ -1,4 +1,4 @@
 import 'module-alias/register'
-import logger from '%src/lib/logger'
+import logger from '%src/tools/logger'
 
 logger.info('Started.')
